htmlscraper/requirements.txt
grpcio
grpcio-tools
parsel